+++ tests/dataset_boston_housing_test.rs
@@ -1,9 +1,9 @@
 #![cfg(feature = "datasets")]
 
 use dataset_core::datasets::boston_housing::*;
-use dataset_core::utils::{download_to, file_sha256_matches, unzip};
+use dataset_core::utils::{download_to, file_sha256_matches};
 use std::fs::File;
-use std::fs::{create_dir_all, remove_dir_all, rename};
+use std::fs::{create_dir_all, remove_dir_all};
 use std::io::Write;
 use std::path::Path;
 
@@ -42,23 +42,10 @@ fn test_boston_housing_no_need_download() {
     // download Boston Housing dataset in advance
     {
         download_to(
-            "https://gist.github.com/nnbphuong/def91b5553736764e8e08f6255390f37/archive/373a856a3c9c1119e34b344de9230ae2ea89569d.zip",
-            download_dir_path
-        ).unwrap();
-
-        // Extract file
-        unzip(
-            &download_dir_path.join("373a856a3c9c1119e34b344de9230ae2ea89569d.zip"),
+            "https://github.com/selva86/datasets/raw/master/BostonHousing.csv",
             download_dir_path,
-        )
-        .unwrap();
-
-        let src = download_dir_path
-            .join("def91b5553736764e8e08f6255390f37-373a856a3c9c1119e34b344de9230ae2ea89569d")
-            .join("BostonHousing.csv");
-        let dst = download_dir_path.join("BostonHousing.csv");
-        // move boston_housing.csv out of the directory
-        rename(src, &dst).unwrap();
+            None,
+        ).unwrap();
     }
 
     // should use cached Boston Housing dataset
@@ -90,7 +77,7 @@ fn test_boston_housing_overwrite() {
     assert!(
         file_sha256_matches(
             &download_dir_path.join("BostonHousing.csv"),
-            "c9aef7e921f2b44d4e7a234aea24f478186d5d457c3758035864b083ac8e7451"
+            "ab16ba38fbbbbcc69fe930aab1293104f1442c8279c130d9eba03dd864bef675"
         )
         .unwrap()
     );