+++ changelog.md
@@ -5,6 +5,12 @@ This change log records updates after 2026-2-13.
 
 Please view [SomeB1oody/dataset-core](https://github.com/SomeB1oody/dataset-core) for more info.
 
+## [0.1.0] - 2026-4-11
+### Changed
+- Replace `downloader` with `ureq` for file downloads
+- split wine quality dataset into red and white datasets
+- refactor `download_to` to provide optional file naming
+
 ## [0.1.0] - 2026-4-10
 ### Added
 - Integrate `thiserror` for structured error handling, and refactor error implementations accordingly